06 se crearon las pestanias inscripciones y asistencias y se agrego a…

+++ src/Clases_utilizadas/ranking.java
@@ -19,7 +19,7 @@ public class ranking {
     }
     public void CalculaPresentismo(){
         if(totclases!=0)
-            Presentismo = asistencias/totclases*100;
+            Presentismo = ( (double)asistencias / totclases * 100 );
         else
             Presentismo = 0;
     }
